Форма удаления готова, начало формы регистрации

+++ AutoSalonn/DelCarForm.cs
@@ -16,29 +16,42 @@ namespace AutoSalonn
         public DelCarForm()
         {
             InitializeComponent();
+            for(int i=0; i<FiltrForm.cars_list.Count; i++)
+            {
+                comboBox1.Items.Add(FiltrForm.cars_list[i].name);
+            }
         }
 
         private void DelButton_Click(object sender, EventArgs e)
         {
-            File.Delete("cars.txt");
-            for(int i=0; i<FiltrForm.cars_list.Count; i++)
+            var result = MessageBox.Show("Вы действительно хотите удалить выбранный объект", "Удаление объекта", MessageBoxButtons.YesNo);
+            if (result == DialogResult.Yes)
             {
-                if(textBox1.Text == FiltrForm.cars_list[i].name)
+                File.Delete("cars.txt");
+                for (int i = 0; i < FiltrForm.cars_list.Count; i++)
                 {
+                    if (textBox1.Text == FiltrForm.cars_list[i].name)
+                    {
 
+                    }
+                    else
+                    {
+                        File.AppendAllText("cars.txt",
+                                        FiltrForm.cars_list[i].name + ", " +
+                                        FiltrForm.cars_list[i].kuzov + ", " +
+                                        FiltrForm.cars_list[i].kpp + ", " +
+                                        FiltrForm.cars_list[i].price +
+                                        Environment.NewLine);
+                    }
                 }
-                else
-                {
-                    File.AppendAllText("cars.txt",
-                                    FiltrForm.cars_list[i].name + ", " +
-                                    FiltrForm.cars_list[i].kuzov + ", " +
-                                    FiltrForm.cars_list[i].kpp + ", " +
-                                    FiltrForm.cars_list[i].price +
-                                    Environment.NewLine);
-                }
-            }
-            MessageBox.Show("Удаление прошло успешно");
-            Close();            
+                MessageBox.Show("Удаление прошло успешно");
+                Close(); 
+            }                       
+        }
+
+        private void comboBox1_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            textBox1.Text = comboBox1.Text;
         }
     }
 }
